requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.32.0
Django==4.1.7
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14